requirements.txt
unsloth
transformers==4.56.2
trl==0.22.2
bitsandbytes
accelerate
peft
triton
cut_cross_entropy
unsloth_zoo
sentencepiece
protobuf
datasets>=3.4.1,<4.0.0
huggingface_hub>=0.34.0
hf_transfer
evaluate
jiwer
torchcodec
librosa
soundfile
python-dotenv
ctranslate2 
